Have user_event and user_event_queue working for kqueue runloop.
The user_event_queue is no longer a variant of the union runloopevent type, and the queuewatcher type has been eleminated.

Using the kqueue EVFILT_USER mechanism events get combined so one needs a mechanism over and above kqueue to make sure
triggered events are not misted - the typical edge-trigger problem. In the case of UserEventQueue the queue itself is the mechanism
to solve this problem as on a user event the code removes things from the queue until the queue is empty.

I have also implemented a dup() of the kqueue fd so that the trigger call, which happens inside a mutext block, does not
limit access to a threads  main kqueue file descriptor.

Finally have starte dimplementing the 2 pipe trick for kqueue.

Next steps:

1. For kqueue runloop - a modified version of UserEvent to implement a semaphore style user event like eventfd
2. For epoll runloop - make the same mods as for kqueue - remove QueueWatcher and make UserEventQueue NOT an instance of RunloopEvent